Allow hashing symbolic link data
When checking the integrity of filesystem-based backups (created by
tools like `cp -al`, `rsync`, or `rsnapshot`) it's desirable to check
the integrity of symbolic links as well, i.e. to check that they still
refer to the same path as when the backup was created.

Add the `--hash-symlink-data` option to hash the data in the symbolic
link itself, not the file it refers to, and to hash its data even when
it refers to a directory.